some fixes, still some gc bugs

+++ src/ll.c
@@ -3,7 +3,7 @@
 void * must_malloc(size_t bytes) {
     void * ptr = malloc(bytes);
     if (ptr == nil) {
-        panic("Failed to allocate %u bytes", bytes);
+        panic("Failed to allocate %u bytes", (unsigned int) bytes);
     }
     return ptr;
 }
